Move files to email-templates plugin
