+++ src/plugins/email-templates/mutations/index.js
@@ -1,5 +1,7 @@
 import renderEmail from "./renderEmail.js";
+import updateTemplate from "./updateTemplate.js";
 
 export default {
-  renderEmail
+  renderEmail,
+  updateTemplate
 };
