+++ src/plugins/email-templates/simpleSchemas.js
